Removed code duplication. Added support for skipping special characters like quotes, double quotes and other.
